Platform/RaspberryPi/Acpitables: Add eMMC2 device and tweak Arasan
(cherry picked from commit 311ed3c79047e2a7ad001ab94b240ac6558bf06c)

The primary problem with the RPi's Arasan controller is
the lack of a meaningful capabilities register. With just
a sdhci-caps _DSD entry we can provide that information. It
can then be bound to the Linux sdhci_iproc driver which
already hardcodes the remaining controller bugs.

Further we have gotten BRCME88C approved as the HID
for the newer eMMC2 controller. So lets define an
ACPI object to describe it.

Of course both devices are sharing an interrupt so
we should also indicate that in the table as well.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>
Reviewed-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Andrei Warkentin <awarkentin@vmware.com>
(cherry picked from commit 311ed3c79047e2a7ad001ab94b240ac6558bf06c)

Change-Id: Ieea450096bec53d47d0f2a55a1477d37a5914d15