Platform/RaspberryPi/Acpitables: Correct duplicate _UIDs
(cherry picked from commit d22356576053271b2645afa1387ed909050027ae)

The new emmc2 device adds a container with a _UID of 0 which duplicates
an existing ACPI container.

This causes windows to fail to boot with a Stop ACPI_BIOS_ERROR.  The
workaround without this fix is to go into the BDS menus and set the SD
routing to the Arasan.

Fixes: 311ed3c790 ("Platform/RaspberryPi/Acpitables: Add eMMC2 device and tweak Arasan")
Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>
Reviewed-by: Ard Biesheuvel <ardb@kernel.org>
(cherry picked from commit d22356576053271b2645afa1387ed909050027ae)

Change-Id: I02beabfa6cb909f5e489726600f7c7c95d92c870