MinPlatformPkg/TestPointCheckLib: Improve adjacent region checking
(cherry picked from commit 5021f0151728ca85772bb4424c2518b267a812e0)

REF:https://bugzilla.tianocore.org/show_bug.cgi?id=3521

The current logic depends on a particular order in which the
descriptors for three or more regions are placed in the array to
perform proper adjacency checking. When three or more regions are
all adjacent, but neighboring descriptors are not adjacent, the
logic can improperly report a failure. Adjust the logic so that
all descriptors are checked for adjacency.

Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Eric Dong <eric.dong@intel.com>
Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
(cherry picked from commit 5021f0151728ca85772bb4424c2518b267a812e0)

Change-Id: I4e6c5a1622a7cd6adaab6202e7c7dc05f1eeadbe