Silicon/Qemu: don't advertise GICC legacy mmio interface in SbsaQamu MADT
(cherry picked from commit 69035742d336703497f178ef8d273d9ed56c7d0b)

The MADT GICC structure contains the field PhysicalBaseAddress, which
is needed for a GICv1/v2 implementation, or to indicate legacy
compatibility in modern GICs.

Linux commit 9739f6ef053f1, included in v5.12, adds a warning message
when this field is populated but invalid:
  [Firmware Bug]: CPU interface incapable of MMIO access

As it happens, we currently initialize this to PcdGicDistributorBase
instead of PcdGicInterruptInterfaceBase, and as a result we now trigger
this warning.

Since this is an SBSA reference implementation, and legacy GIC support
has never worked for this port, set the field to 0.

Signed-off-by: Leif Lindholm <leif@nuviainc.com>
Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Graeme Gregory <graeme@nuviainc.com>
Cc: Radoslaw Biernacki <rad@semihalf.com>
Acked-by: Ard Biesheuvel <ardb@kernel.org>
(cherry picked from commit 69035742d336703497f178ef8d273d9ed56c7d0b)

Change-Id: Id54d5ee16cb982043fe4a5275e716b28fd288990