Platform/ARM/Morello: Enable SoC platform I2C HDMI bus and GOP TDA19988
Enable platform configuration of the I2C bus used to configure the
HDMI output related chips.
The TDA19988 on this bus is connected to the DPU so when the TDA19988
driver have been probed and installed it is marked as the
Morello Platform GOP TDA19988.

Signed-off-by: Varshit Pandya <varshit.pandya@arm.com>
(cherry picked from commit 000263c24fee9fd861be696ac699c7e37544f946)