Platform/BeagleBoard: Drop PeCoff protocol
(cherry picked from commit b7505f9c27f11064373ce3359ec312f48b265643)

BeagleBoard is the only user of the so-called 'PE/COFF protocol', which
just exposes the PE/COFF loader library API via a protocol, presumably
to avoid duplicating this code in the PrePi SEC component as well as the
DXE core.

This is a rather questionable practice, and it would be better to drop
this code so we can remove it from the EDK2 main repo as well.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
(cherry picked from commit b7505f9c27f11064373ce3359ec312f48b265643)

Change-Id: I724451a1451532cd19e324548184fa50621a2620