Platform/Sgi: Add SSDT table for Virtio-P9
(cherry picked from commit 70e97c37ae294a3882684cfe8346bf89fd00ae8e)

Some of the Arm reference design FVP platforms support the Virtio-p9
device as part of the RoS subsystem. Add an entry for this device in
the SSDT acpi table.

The device entry is listed in a new SSDT file as only some of the
reference design FVP platforms support it and so this file is included
in the build for only the platforms that support Virtio-P9 device.

Signed-off-by: Vivek Gautam <vivek.gautam@arm.com>
Reviewed-by: Pierre Gondois <pierre.gondois@arm.com>
Reviewed-by: Sami Mujawar <sami.mujawar@arm.com>
(cherry picked from commit 70e97c37ae294a3882684cfe8346bf89fd00ae8e)

Change-Id: I2f8bcdbeca53c126fd37491b09a7d270aadbbfe2