Platform/ARM/JunoPkg: Bump the FV size to 2MB
(cherry picked from commit e702edbb394fad404d82646fac429903d5d67e70)

Increase the FV size from 996KB to 2MB. This allows more features to be
accommodated in the firmware image, such as HTTPS boot. It also allows a
NOOPT build to be added, which is useful for source-level debugging.

Add a comment explaining that the size chosen doesn't reflect the actual
size of the Flash Device but is instead selected to reduce the amount of
time needed to copy the image to the target.

Update the BlockSize and NumBlocks to reflect the 256KB block size of
the main area on the Juno R2.

Signed-off-by: Rebecca Cran <rebecca@bsdio.com>
(cherry picked from commit e702edbb394fad404d82646fac429903d5d67e70)

Change-Id: Iae55234f1fbac4ac0dc4f246e483ed066a9e281b