Platform/Sgi: Add a new PCD for L2 cache size
(cherry picked from commit 215928f5f12aa910175175217f128697836b5381)

A new PCD named PcdL2CacheSize is introduced to specify the size of the
CPU's L2 cache. This PCD can be used in PPTT tables for a reference
design platform. The default value of this PCD is set to 1MB but can be
changed depending on the variant of the reference design platform.

Signed-off-by: Prabin CA <prabin.ca@arm.com>
Reviewed-by: levi.yun <yeoreum.yun@arm.com>
Reviewed-by: Sami Mujawar <sami.mujawar@arm.com>
(cherry picked from commit 215928f5f12aa910175175217f128697836b5381)

Change-Id: Id9aa63673f10455269ffa111a0e6e5116e416738