Platform/RaspberryPi/PlatformBootManagerLib: Reimplement reset hook
(cherry picked from commit c742472e701a5aad9394302b7447ba78492aeecb)

Duplicate the logic that is triggered on a system reset into the
platform boot manager driver, and hook it up to the EDK2 platform
specific reset notification driver. This is supported by generic EDK2
code in MdeModulePkg, allowing us to retire the platform-specific
EfiResetSystemLib implementation in a subsequent patch. This is needed
because this library class and its only user ResetRuntimeDxe in
EmbeddedPkg are deprecated and are going to be removed.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
(cherry picked from commit c742472e701a5aad9394302b7447ba78492aeecb)

Change-Id: I7254d2dba16570650be66f6864c89dbf6354f3e1