Platform/RaspberryPi: Use depex based dispatch order for varstore
(cherry picked from commit f9a8aca307180e134a3365b50e030765b01901bf)

The VarBlockServiceDxe driver needs to be dispatched before the common
VariableRuntimeDxe, but we are currently relying on FDF order and lack
of transitive dependencies for this, which is fragile, and will break
once we move to the generic reset runtime.

So use the existing helper library for this, which can be plugged into
the generic variable drivers, and force them to depex on a GUID that can
be installed as a NULL protocol in VarBlockServiceDxe.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Reviewed-by: Leif Lindholm <quic_llindhol@quicinc.com>
(cherry picked from commit f9a8aca307180e134a3365b50e030765b01901bf)

Change-Id: I1e5dada45acf00e441f1253edc1cfb5c319d5256