Platform/ARM/VExpressPkg: Drop FVP dependency on ArmGicArchLib
(cherry picked from commit 98c8f583849e6124d66dd456aef22125395463b4)

Instead of using ArmGicArchLib to decide whether or not the GIC is in v3
mode (and inadvertently switching it into that mode behind the back of
the GicDxe driver if it wasn't already), use the presence of the GIC
sysreg interface to infer this. This is needed because ArmGicArchLib is
going to be removed entirely.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
(cherry picked from commit 98c8f583849e6124d66dd456aef22125395463b4)

Change-Id: Idb0734b06ed2dc303ffdff9a1ddb16423d2345e4